Increase fidelity of Game Score
10ths or even 100ths of a second may matter to someone... So, it's best
to start collecting this level of fidelity now, before it's too late!

+++ db/migrate/20240809213612_create_games.rb
@@ -18,7 +18,7 @@ class CreateGames < ActiveRecord::Migration[7.1]
       t.string(:type, null: false)
       t.datetime(:started_at)
       t.datetime(:ended_at, index: true)
-      t.integer(:score, index: true)
+      t.float(:score, index: true)
 
       t.timestamps
       t.index(:created_at)
